+++ src/files.rs
@@ -26,9 +26,9 @@ pub struct File {
   metadata: HashMap<String, String>,
   asset_ids  : Vec<u64>,
   uploaded  : bool,
-  uploaded_at  : u128,
-  created_time : u128,
-  last_updated_time : u128
+  uploaded_at  : u64,
+  created_time : u64,
+  last_updated_time : u64
 }
 
 pub struct Files {
