+++ src/time_series.rs
@@ -26,8 +26,8 @@ pub struct TimeSerie {
   description: String,
   security_categories: Vec<u64>,
   id: u64,
-  created_time: u128,
-  last_updated_time: u128
+  created_time: u64,
+  last_updated_time: u64
 }
 
 pub struct TimeSeries {
